refactor:move themed text to ui
