+++ mobile/src/app/+not-found.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link, Stack } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/components/ui';
 import { ThemedView } from '@/components/ThemedView';
 
 export default function NotFoundScreen() {
