+++ mobile/src/components/reuseables/TitledScreen.tsx
@@ -1,7 +1,7 @@
-import { StyleProp, StyleSheet, Text, TextStyle, View, ViewStyle } from 'react-native';
+import { StyleProp, StyleSheet, TextStyle, ViewStyle } from 'react-native';
 import React from 'react';
 import { ThemedView } from '@/components/ThemedView';
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/components/ui';
 import { globals } from '@/styles/globals';
 
 const TitledScreen = ({
