+++ mobile/src/components/reuseables/index.ts
@@ -1,1 +1,1 @@
-export { default as BottomSheet } from './BottomSheet';
+export { default as BottomSheet } from '../ui/BottomSheet';
