+++ mobile/src/components/ui/index.ts
@@ -1,2 +1,5 @@
 export { default as Button } from './Button';
 export { default as Pill } from './Pill';
+export { default as EmptyState } from './EmptyState';
+export { default as BottomSheet } from './BottomSheet';
+export { default as ThemedText } from './ThemedText';
