+++ mobile/src/components/ui/settings/DefaultTabSheet.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { StyleSheet, TouchableOpacity, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import { BottomSheetModal } from '@gorhom/bottom-sheet';
 import { FontAwesome, Ionicons } from '@expo/vector-icons';
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/components/ui';
 import { useTheme } from '@/hooks/useTheme';
 import { type Tab } from '@/hooks/usePreferences';
 import { tabs } from '@/constants/data';
-import { BottomSheet } from '@/components/reuseables';
+import { BottomSheet } from '@/components/ui';
 import { HapticButton } from '@/components/layout';
 
 type DefaultTabSheetProps = {
